paramCSbackend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
colorama==0.4.6
distlib==0.3.9
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
django-mssql-backend==2.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
filelock==3.16.1
iniconfig==2.0.0
mssql-django==1.5
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pip==24.3.1
pipenv==2024.3.1
platformdirs==4.3.6
pluggy==1.5.0
PyJWT==2.9.0
pyodb==0.1.5
pyodbc==5.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.27.1